refactor toggles menu code
