+++ src/components/menu/TogglesMenu/styles.js
@@ -1,0 +1,5 @@
+import { makeStyles } from "@material-ui/core/styles";
+import styled from "styled-components";
+import List from "@material-ui/core/List";
+
+export const StyledList = styled(List)``;
