+++ src/components/menu/menu.js
@@ -1,11 +1,5 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
-import ListItemText from "@material-ui/core/ListItemText";
-import Switch from "@material-ui/core/Switch";
-import Drawer from "@material-ui/core/Drawer";
 import Fab from "@material-ui/core/Fab";
 import MenuIcon from "@material-ui/icons/Menu";
 import EditIcon from "@material-ui/icons/Edit";
@@ -17,12 +11,11 @@ import ABMSubmenu from "./ABMSubmenu/ABMSubmenu";
 import TypeMenuContainer from "./TypeMenuComponents/TypeMenuContainer";
 import NearMeIcon from "@material-ui/icons/NearMe";
 import NavigationIcon from "@material-ui/icons/Navigation";
+import TogglesMenu from "./TogglesMenu";
 
 // import CircularProgress from "@material-ui/core/CircularProgress";
 
 function Menu(props) {
-    const togglesMeta = settings.menu.toggles;
-    const listOfToggles = Object.keys(togglesMeta);
     const useStyles = makeStyles(theme => ({
         root: {
             width: "100%",
@@ -88,14 +81,6 @@ function Menu(props) {
         setState({ ...state, [side]: open });
     };
 
-    const sideList = side => (
-        <div
-            className={classes.list}
-            role="presentation"
-            onClick={toggleDrawer(side, false)}
-            onKeyDown={toggleDrawer(side, false)}
-        ></div>
-    );
     const handleToggle = value => () => {
         const currentIndex = toggleStateArray.indexOf(value);
         const newChecked = [...toggleStateArray];
@@ -110,61 +95,15 @@ function Menu(props) {
         props.listenToMenuUI(newChecked);
     };
 
-    /**
-     * gets props with initial menu state
-     * and turn on the layer on init
-     */
-    let togglesCompsArray = [];
-    // array of loaded API modules
-    const loadedModlues = Object.keys(props.cityIOdata);
-    // create each toggle
-    for (let i = 0; i < listOfToggles.length; i++) {
-        // check if the mdoule of this toggle
-        // was loaded on the API
-        let requireModule = togglesMeta[listOfToggles[i]].requireModule;
-
-        if (loadedModlues.includes(requireModule) || requireModule === false) {
-            const thisToggle = (
-                <ListItem key={listOfToggles[i]}>
-                    <ListItemText
-                        primary={togglesMeta[listOfToggles[i]].displayName}
-                    />
-                    <ListItemSecondaryAction>
-                        <Switch
-                            edge="end"
-                            onChange={handleToggle(listOfToggles[i])}
-                            checked={
-                                props.menuState.includes(listOfToggles[i])
-                                    ? true
-                                    : false
-                            }
-                        />
-                    </ListItemSecondaryAction>
-                </ListItem>
-            );
-            togglesCompsArray.push(thisToggle);
-        }
-    }
-
     return (
         <>
             <div className={classes.root}>
-                <Drawer
-                    BackdropProps={{
-                        classes: {
-                            root: classes.backDrop
-                        }
-                    }}
-                    anchor="left"
+                <TogglesMenu
+                    classes={classes}
                     open={state.left}
-                    onClose={toggleDrawer("left", false)}
-                >
-                    {sideList("left")}
-                    <List className={classes.root}>
-                        <h2>cityscopeJS</h2>
-                        {togglesCompsArray}
-                    </List>
-                </Drawer>
+                    toggleDrawer={toggleDrawer}
+                    handleToggle={handleToggle}
+                />
             </div>
             <Fab
                 className={classes.menuButton}
